ruthlessly remove duplicate functions
The NaCl codebase comes from the SUPERCOP benchmark suite, in which it
is useful to have multiple names for the same function:
generic (crypto_stream) and algorithm-specific (crypto_stream_xsalsa20).
SUPERCOP goes further, with multiple implementations (some of which
perform faster on certain architectures).

Tweetnacl unfortunately kept some of that complexity (to maintain
compatibility with the regular NaCl API).

We don't need that here. crypto_secretbox() means XSalsa20/Poly1305, and
we don't need a crypto_secretbox_xsalsa20poly1305() to remind us of
that. We certainly don't need both functions. So the details of which
algorithm are being used has been moved into a docstring, and only the
generic function has been retained.

This is probably only visible in two places:

* crypto_hash() means SHA512. This is the one function name that might
  be ambiguous.
* crypto_stream() means XSalsa20 (with the 24-byte nonce). It is no
  longer possible to access crypto_stream_salsa20 (8-byte nonce)